+++ example/src/app/flat/flat.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit, HostBinding } from '@angular/core';
+import { mockItems, mockGroups } from '../examples/mocks';
+import { GanttViewType } from 'ngx-gantt';
+
+@Component({
+    selector: 'example-flat',
+    templateUrl: './flat.component.html',
+    styleUrls: ['./flat.component.scss']
+})
+export class AppFlatExampleComponent implements OnInit {
+    constructor() {}
+
+    items = mockItems;
+
+    groups = mockGroups;
+
+    options = {
+        viewType: GanttViewType.month,
+        draggable: true,
+        mergeIntervalDays: 3
+    };
+
+    @HostBinding('class.gantt-demo') class = true;
+
+    ngOnInit(): void {}
+}