+++ packages/gantt/src/flat/examples/basic/basic.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-flat-basic-example',
+    templateUrl: './basic.component.html',
+})
+export class NgxGanttFlatBasicExampleComponent implements OnInit {
+    constructor() { }
+
+    ngOnInit(): void { }
+}